hemoglobin_backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning & Deep Learning
tensorflow==2.15.0
keras==2.15.0
scikit-learn==1.3.2
scikit-image==0.22.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
imageio==2.33.0

# Feature Extraction
mahotas==1.4.13
pywavelets==1.5.0

# Explainability
shap==0.44.0
tf-keras-vis==0.8.5

# Report Generation
reportlab==4.0.7
matplotlib==3.8.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utilities
numpy==1.26.2
pandas==2.1.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
